--- BeerSender.Domain/Boxes/BoxCreated.cs
@@ -1,3 +1,0 @@
-﻿namespace BeerSender.Domain.Boxes;
-
-public record BoxCreated(BoxCapacity Capacity);
