--- BeerSender.Domain/Boxes/CreateBox.cs
@@ -1,5 +1,0 @@
-﻿namespace BeerSender.Domain.Boxes;
-
-public record CreateBox(
-    Guid BoxId,
-    int DesiredNumberOfSpots);
