--- BeerSender.Domain/Boxes/ShippingLabelAdded.cs
@@ -1,3 +1,0 @@
-﻿namespace BeerSender.Domain.Boxes;
-
-public record ShippingLabelAdded(ShippingLabel Label);
