+++ BeerSender.Domain/Events/BoxCreated.cs
@@ -1,0 +1,5 @@
+﻿using BeerSender.Domain.Model;
+
+namespace BeerSender.Domain.Events;
+
+public record BoxCreated(BoxCapacity Capacity);
