+++ BeerSender.Domain/Events/ShippingLabelAdded.cs
@@ -1,0 +1,5 @@
+﻿using BeerSender.Domain.Model;
+
+namespace BeerSender.Domain.Events;
+
+public record ShippingLabelAdded(ShippingLabel Label);
